Creating BasicAuxCrudController and adding some documentation to endpoints
